+++ lib/cubdb.ex
@@ -750,6 +750,27 @@ defmodule CubDB do
     GenServer.call(db, :dirt_factor, :infinity)
   end
 
+  @spec writes_since_compaction(server) :: non_neg_integer
+
+  @doc """
+  Returns the number of writes performed since the last compaction.
+
+  This number is reset to 0 after a compaction operation, and increases by one
+  on each write operation on the BTree. Deletions are considered writes, as they
+  modify the BTree. The way writes are computed depends on the specific BTree
+  implementation, and might not map one-to-one to the high level write
+  operations performed by the user. This metric is provided only as a hint in
+  case of manual compaction, where it is often a good idea to skip a wasteful
+  compaction if only a few write operations were performed since the last
+    compaction.
+
+    The number of writes does not include uncommitted operations, therefore old
+    halted transactions that are never committed are not counted.
+  """
+  def writes_since_compaction(db) do
+    GenServer.call(db, :writes_since_compaction, :infinity)
+  end
+
   @spec put(server, key, value) :: :ok
 
   @doc """
@@ -1304,6 +1325,13 @@ defmodule CubDB do
 
   def handle_call(:dirt_factor, _, state = %State{btree: btree}) do
     {:reply, Btree.dirt_factor(btree), state}
+  end
+
+  def handle_call(:writes_since_compaction, _, state) do
+    %State{btree: btree} = state
+    %Btree{dirt: dirt} = btree
+
+    {:reply, dirt, state}
   end
 
   def handle_call(:start_transaction, from, state = %State{writer: nil}) do