+++ CHANGELOG.md
@@ -6,6 +6,14 @@ releases](https://github.com/lucaong/cubdb/releases).
 Since `v1.0.0`, `CubDB` follows [semantic versioning](https://semver.org), and
 reports changes here.
 
+## Unreleased
+
+  - Add function `CubDB.writes_since_compaction/1` to get the number of writes since the last successful compaction
+
+Bug fixes:
+
+  - Fix dirt calculation upon restart of `CubDB`
+
 ## v2.0.2 (2023-01-01)
 
 Bug fixes: